Some renaming as per new agreement.
