+++ src/app/[lang]/admin/crags/[cragSlug]/sectors/[sectorId]/routes/lib/delete-route-action.ts
