+++ src/app/[lang]/admin/crags/[cragSlug]/sectors/components/delete-sector-dialog.tsx
@@ -1,7 +1,7 @@
 import Dialog from "@/components/ui/dialog";
 import { Sector } from "@/graphql/generated";
 import { Dispatch, SetStateAction, useState } from "react";
-import deleteSectorAction from "../server-actions/delete-sector-action";
+import deleteSectorAction from "../lib/delete-sector-action";
 import { useRouter } from "next/navigation";
 import { labelAndNameToString } from "@/utils/sector-helpers";
