+++ src/app/[lang]/admin/crags/[cragSlug]/sectors/components/edit-sectors-none.tsx
@@ -7,7 +7,7 @@ import { Crag } from "@/graphql/generated";
 import { usePathname, useRouter } from "next/navigation";
 import ConvertToSectorsManyDialog from "./convert-to-sectors-many-dialog";
 import { useState } from "react";
-import createSectorAction from "../server-actions/create-sector-action";
+import createSectorAction from "../lib/create-sector-action";
 import { useAuthContext } from "../../../../../../components/auth-context";
 import { canEdit } from "@/utils/contributables-helpers";
