+++ src/app/[lang]/admin/crags/[cragSlug]/sectors/lib/delete-sector-action.ts
